refactor(printable-itinerary): Covert remaining files to TypeScript.
